requirements.txt
librosa>=0.10.0
soundfile>=0.12.1
numpy>=1.24.0
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
# Optional: for cutout effect
# rembg>=2.0.50
# onnxruntime>=1.16.0
